Add files via upload
In this project I planned on building a project to help me with improving my typing speed. In this any user need to input a word/ string he needs to practise and then need to practise it 5 time in a row as fast as possible.
I deployed mathPlot library to present the user with the time taken by each attempt and number of mistakes he did to help him analyze.